% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalauc-package.R
\docType{package}
\name{survivalauc-package}
\alias{survivalauc}
\alias{survivalauc-package}
\title{survivalauc: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/survivalauc}
  \item Report bugs at \url{https://github.com/tidymodels/survivalauc/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
