% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{compute_auc_components}
\alias{compute_auc_components}
\title{Compute Survival AUC}
\usage{
compute_auc_components(time, status, score, threshold)
}
\arguments{
\item{time}{Numeric, survival/censoring time.}

\item{status}{Numeric, censoring indicator taking value 1 if survival time is
observed and 0 otherwise.}

\item{score}{Numeric, risk score obtained from a risk prediction model.}

\item{threshold}{Numeric, thresholds used in estimating sensitivity and
specificity.}
}
\description{
Implementing the alternative estimators of the time-dependent area
under the ROC curve, time-dependent sensitivity and specificity,
and time-dependent ROC curves.
}
\details{
\code{time}, \code{status}, and \code{score} must have the same length.

The corresponding estimators are presented in equations (9)-(11) in
Chambless and Diao (Statistics in Medicine, 2006; 25: 3474-3486).
}
